[GOBBLIN-2004] Helix-free yarn app launcher
