+++ gobblin-yarn/src/main/java/org/apache/gobblin/yarn/GobblinApplicationMaster.java
@@ -65,6 +65,7 @@ import org.apache.gobblin.util.logs.Log4jConfigurationHelper;
 import org.apache.gobblin.util.logs.LogCopier;
 import org.apache.gobblin.util.reflection.GobblinConstructorUtils;
 import org.apache.gobblin.yarn.event.DelegationTokenUpdatedEvent;
+import org.apache.gobblin.yarn.helix.HelixMessageSubTypes;
 
 
 /**
